Change unit test inputs for loopfilter unit tests
The high bitdepth loopfilter functions are specialized on certain
filtering scenarios where the value of the masks indicates there is no
need to compute all filter sizes. This optimization will also be added
to the standard bit-depth implementation in future commits.

The inputs of the unit test cover only a very reduced set of
these scenarios, so use the function from the libaom loopfilter unit
tests to generate inputs that will cover all filtering cases.