cleanup: bump version to v3.0.1
I forgot to bump the version numbers in the repo before cutting the
release. Unfortunately, this means that the release will still point to
the original commit, however, this shouldn't really have a large effect
as there were no major api changes to the code that would require
checking for 3.0.1 specifically.

Signed-off-by: Christopher Degawa <ccom@randomderp.com>